==> 04 Prediction model.Rmd <==
---
title: "Prediction_model"
date: "11/7/2020"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
library(gtools)
library(lubridate)
library(caret)
library(geosphere)
library(xml2)
library(rvest)
library(curl)
```

## Load dataset prepared for prediction. 

### Chosen predictors will be Property type, Size, low/high floor, age of hdb, HDB Town,  no. bus stop, min pri school and secondary school dist, min mall and supermarket dist, min mrt distance

```{r load dataset}
dat <- read.csv("D:/Year4_SEM1/DBA3702_R/Proj/dat_for_predict2.csv")

dat_Pred <- subset(dat, select = -c(long,lat,Groceries...Supermarts,Shopping.Malls,Secondary.Schools,Primary.Schools,Bus.Stops,Trains..MRT...LRT.))

dat_Pred$Property.Type <- replace(dat_Pred$Property.Type, dat_Pred$Property.Type==1,"HDB 1 Room")
dat_Pred$Property.Type <- replace(dat_Pred$Property.Type, dat_Pred$Property.Type==2,"HDB 2 Rooms")
dat_Pred$Property.Type <- replace(dat_Pred$Property.Type, dat_Pred$Property.Type==3,"HDB 3 Rooms")
dat_Pred$Property.Type <- replace(dat_Pred$Property.Type, dat_Pred$Property.Type==4,"HDB 4 Rooms")
dat_Pred$Property.Type <- replace(dat_Pred$Property.Type, dat_Pred$Property.Type==5,"HDB 5 Rooms")
dat_Pred$Property.Type <- replace(dat_Pred$Property.Type, dat_Pred$Property.Type==6,"HDB Executive")
dat_Pred$Property.Type <- replace(dat_Pred$Property.Type, dat_Pred$Property.Type==7,"HDB Jumbo")

#Create dummy for property type
dmy <- dummyVars(" ~ Property.Type", data = dat_Pred)
trsf <- data.frame(predict(dmy, newdata = dat_Pred))
dat_Pred <- cbind(dat_Pred,trsf)

#Remove and set 4 rooms and amk town as reference group
dat_Pred <- subset(dat_Pred,select = -c(Property.Type,`Property.TypeHDB.4.Rooms`, HDB.TownAng.Mo.Kio))
head(dat_Pred)
```

# 1. Resale
## --------------------
## Predict model 1 

```{r predict model 1}
set.seed(3456)
samp <- createDataPartition(dat_Pred$Asking, p = 0.8, list = FALSE)
training <- dat_Pred[samp,]
testing <- dat_Pred[-samp,]

#Predict model
mod <- lm(Asking ~ . -X, data = training)
print(summary(mod))
```

### From the linear regression, it seems like Serangoon, bedok, number of bus stops, supermarket dist, park dist, clinic dist, 1 room and 2 rooms hdb is not significant. So we will remove these variable.

```{r predict model 2}
mod2 <- lm(Asking ~ . -X-HDB.TownBedok-HDB.TownSerangoon-no.bus_stops-min_supermarket_dist-min_park_dist-min_clinic_dist-Property.TypeHDB.1.Room-Property.TypeHDB.2.Rooms , data = training)

print(summary(mod2))

#Predict value using model 2 and add to dataset 
dat_Pred2 <- dat_Pred
dat_Pred2$predict_value <- predict(mod2, dat_Pred)

print(head(dat_Pred2))
#write.csv(dat_Pred2,"D:\\Year4_SEM1\\DBA3702_R\\Proj\\predicted_HDB2.csv", row.names = FALSE)
```

## Predict future resale price in 5 years (Resale Flats)

```{r predict future resale}
dat_Pred_future <- read.csv("D:/Year4_SEM1/DBA3702_R/Proj/predicted_HDB2.csv")

#Remove 1 room, serangoon, supermarket dist as data too little (also not significant at 0.5)
mod2 <- lm(Asking ~ . -X-predict_value-HDB.TownSerangoon-min_supermarket_dist-Property.TypeHDB.1.Room-min_park_dist-min_clinic_dist-Property.TypeHDB.2.Rooms, data = dat_Pred_future)

summary(mod2)
```

### Increase age by 5 & get Bala table (housing discount rate based on age)

```{r age and bala}
#Edit Age by adding 5 more years (predicting 5 years later)
dat_Pred_future$Age <- dat_Pred_future$Age+5

dat_Pred_future$predict_value <- predict(mod2, dat_Pred_future)

#Get bala data
url <- "https://stackedhomes.com/editorial/freehold-vs-leasehold-part-1/#gs.jhorb9"
page <- read_html(url)
#closeAllConnections()

df <- as.data.frame(html_table(page))
head(df)
#write.csv(df,"D:\\Year4_SEM1\\DBA3702_R\\Proj\\bala_tab.csv", row.names = FALSE)
```

### Adjust current price to get future housing price (using bala and inflation)

```{r get future price}
dat_Pred_future$lease_left <- 99 - dat_Pred_future$Age
bala_tab <- read.csv("D:/Year4_SEM1/DBA3702_R/Proj/bala_tab2.csv")

#calculate future value rate (after 5 years)
precent_value_future <- bala_tab$Percentage.....of.Freehold.Value[match(dat_Pred_future$lease_left, bala_tab$Term.of.Years)]


#calculate current value rate (now)
dat_Pred_future$lease_left2 <- 99 - dat_Pred_future$Age + 5
precent_value_old <- bala_tab$Percentage.....of.Freehold.Value[match(dat_Pred_future$lease_left2, bala_tab$Term.of.Years)]

discount_rate <- (precent_value_old-precent_value_future)/precent_value_old
dat_Pred_future$discount_rate_bala <- 1+discount_rate

#Add in inflation rate (5 years predicted)
inflation_rate <- 1+0.01459

dat_Pred_future$predict_value_adj <- (dat_Pred_future$predict_value/dat_Pred_future$discount_rate_bala)*inflation_rate
head(dat_Pred_future)
```

### Merge data to hdb_resale (add in predicted prices)
```{r merge}
predict_hdb <- read.csv("D:/Year4_SEM1/DBA3702_R/Proj/predicted_HDB2.csv")
predict_hdb$predict_value_5Y <- dat_Pred_future$predict_value_adj

dat_original <- read.csv('D:/Year4_SEM1/DBA3702_R/Proj/app/hdb_resale.csv')

dat_merge <- merge(dat_original,y=predict_hdb[,c("X","Age","predict_value","predict_value_5Y","Size_sqm")], by="X")

head(dat_merge)
#write.csv(dat_merge,"D:\\Year4_SEM1\\DBA3702_R\\Proj\\resale_merge.csv", row.names = FALSE)
```

## Save Prediction Model as excel (for interpretation)
```{r save model}
dat_Pred_future <- read.csv("D:/Year4_SEM1/DBA3702_R/Proj/predicted_HDB2.csv")

# Convert KM to Meters (for easier interpretation of model)
dat_Pred_future$min_PRIschool_dist <- dat_Pred_future$min_PRIschool_dist*1000
dat_Pred_future$min_SECschool_dist <- dat_Pred_future$min_SECschool_dist*1000
dat_Pred_future$min_malls_dist <- dat_Pred_future$min_malls_dist*1000
dat_Pred_future$min_supermarket_dist <- dat_Pred_future$min_supermarket_dist*1000

#Remove 1 room, serangoon, supermarket dist as data too little (not significant at 0.5)
mod2 <- lm(Asking ~ . -X-predict_value-HDB.TownSerangoon-min_supermarket_dist-Property.TypeHDB.1.Room-min_park_dist-min_clinic_dist-Property.TypeHDB.2.Rooms, data = dat_Pred_future)

model_res <- summary(mod2)
model_summary <- coefficients(model_res)

print(model_summary)
#write.csv(model_summary,"D:\\Year4_SEM1\\DBA3702_R\\Proj\\model_summary_update.csv", row.names = TRUE)
```

# 2. BTO
## -------------
## Using the model to predict for BTO

```{r build model for BTO}
resale_pred <- read.csv("D:/Year4_SEM1/DBA3702_R/Proj/resaleHDB_pred.csv")
resale_pred <- subset(resale_pred,select=-c(predict_value,predict_value_5Y))

resale_pred <- transform(resale_pred, min_school_dist = pmin(min_PRIschool_dist, min_SECschool_dist))
resale_pred <- subset(resale_pred, select = -c(min_PRIschool_dist, min_SECschool_dist))

#Remove 2 rooms and 5 rooms as not statistically significant
resale_pred_new <- subset(resale_pred, select = c(Size_sqm,min_MRT_dist,min_malls_dist,min_park_dist,min_clinic_dist,min_school_dist,Asking,HDB.TownBishan,HDB.TownChoa.Chu.Kang,HDB.TownGeylang,HDB.TownPasir.Ris,HDB.TownSembawang,HDB.TownTampines,HDB.TownToa.Payoh,HDB.TownWoodlands,Property.TypeHDB.3.Rooms,Age))

mod <- lm(Asking ~ ., data = resale_pred_new)

summary(mod)
```
## Clean BTO to fit model

```{r clean BTO 1}
bto <- read.csv("D:/Year4_SEM1/DBA3702_R/Proj/cleaned_data/bto_clean2.csv")
bto$Asking <- (bto$Min.Price + bto$Max.Price)/2

bto <- subset(bto, select=-c(Min.Price,Max.Price)) 
bto <- rename(bto, "HDB.Town" = "Location")
bto <- rename(bto, "Property.Type" = "Property.Type.")
bto <- rename(bto, "Size_sqm" = "Floor.size")
bto <- rename(bto, "min_park_dist" = "min_parks_dist")
bto <- rename(bto, "min_clinic_dist" = "min_clinics_dist")
bto <- rename(bto, "min_school_dist" = "min_schools_dist")

bto$Age <- 5

#HDB town
dmy <- dummyVars(" ~ HDB.Town", data = bto)
trsf <- data.frame(predict(dmy, newdata = bto))
bto <- cbind(bto,trsf)

#remove HDB town
bto <- subset(bto,select=-HDB.Town)

#Property type
dmy <- dummyVars(" ~ Property.Type", data = bto)
trsf <- data.frame(predict(dmy, newdata = bto))
bto <- cbind(bto,trsf)

#remove HDB town
bto <- subset(bto,select=-Property.Type)
head(bto)
```

## Predict BTO 5 years price

```{r predict}
bto$predict_value5Y <- predict(mod, bto)

#Remove 2 rooms flexi as not much data point (inaccurate prediction)
bto$predict_value5Y <- ifelse(bto$Property.TypeHDB.2.Rooms==1,NA,bto$predict_value5Y)

#write.csv(bto,"D:\\Year4_SEM1\\DBA3702_R\\Proj\\predicted_BTO.csv", row.names = FALSE)

bto_predict <- read.csv("D:/Year4_SEM1/DBA3702_R/Proj/predicted_BTO.csv")
bto_original <- read.csv("D:/Year4_SEM1/DBA3702_R/Proj/cleaned_data/bto_clean2.csv")

dat_merge <- merge(bto_original,y=bto_predict[,c("X","Asking","predict_value5Y")], by="X")
head(dat_merge)

#write.csv(dat_merge,"D:\\Year4_SEM1\\DBA3702_R\\Proj\\bto_merge.csv", row.names = FALSE)
```
